+++ com.revolsys.open.core/src/test/java/com/revolsys/jts/testold/algorithm/CGAlgorithmsTest.java
@@ -17,36 +17,36 @@ public class CGAlgorithmsTest extends TestCase {
   }
 
   public void testDistanceLineLineDisjointCollinear() {
-    assertEquals(1.999699, CGAlgorithms.distanceLineLine(new Coordinate(0.0, 0,
-      Coordinates.NULL_ORDINATE), new Coordinate(9.9, 1.4,
-      Coordinates.NULL_ORDINATE), new Coordinate(11.88, 1.68,
-      Coordinates.NULL_ORDINATE), new Coordinate(21.78, 3.08,
+    assertEquals(1.999699, CGAlgorithms.distanceLineLine(new Coordinate((double)0.0, 0,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)9.9, 1.4,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)11.88, 1.68,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)21.78, 3.08,
       Coordinates.NULL_ORDINATE)), 0.000001);
   }
 
   public void testDistancePointLine() {
-    assertEquals(0.5, CGAlgorithms.distancePointLine(new Coordinate(0.5, 0.5,
-      Coordinates.NULL_ORDINATE), new Coordinate(0.0, 0,
-      Coordinates.NULL_ORDINATE), new Coordinate(1.0, 0,
+    assertEquals(0.5, CGAlgorithms.distancePointLine(new Coordinate((double)0.5, 0.5,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)0.0, 0,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)1.0, 0,
       Coordinates.NULL_ORDINATE)), 0.000001);
-    assertEquals(1.0, CGAlgorithms.distancePointLine(new Coordinate(2.0, 0,
-      Coordinates.NULL_ORDINATE), new Coordinate(0.0, 0,
-      Coordinates.NULL_ORDINATE), new Coordinate(1.0, 0,
+    assertEquals(1.0, CGAlgorithms.distancePointLine(new Coordinate((double)2.0, 0,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)0.0, 0,
+      Coordinates.NULL_ORDINATE), new Coordinate((double)1.0, 0,
       Coordinates.NULL_ORDINATE)), 0.000001);
   }
 
   public void testDistancePointLinePerpendicular() {
     assertEquals(0.5, CGAlgorithms.distancePointLinePerpendicular(
-      new Coordinate(0.5, 0.5, Coordinates.NULL_ORDINATE), new Coordinate(0.0,
-        0, Coordinates.NULL_ORDINATE), new Coordinate(1.0, 0,
+      new Coordinate((double)0.5, 0.5, Coordinates.NULL_ORDINATE), new Coordinate((double)0.0,
+        0, Coordinates.NULL_ORDINATE), new Coordinate((double)1.0, 0,
         Coordinates.NULL_ORDINATE)), 0.000001);
     assertEquals(0.5, CGAlgorithms.distancePointLinePerpendicular(
-      new Coordinate(3.5, 0.5, Coordinates.NULL_ORDINATE), new Coordinate(0.0,
-        0, Coordinates.NULL_ORDINATE), new Coordinate(1.0, 0,
+      new Coordinate((double)3.5, 0.5, Coordinates.NULL_ORDINATE), new Coordinate((double)0.0,
+        0, Coordinates.NULL_ORDINATE), new Coordinate((double)1.0, 0,
         Coordinates.NULL_ORDINATE)), 0.000001);
     assertEquals(0.707106, CGAlgorithms.distancePointLinePerpendicular(
-      new Coordinate(1.0, 0, Coordinates.NULL_ORDINATE), new Coordinate(0.0, 0,
-        Coordinates.NULL_ORDINATE), new Coordinate(1.0, 1,
+      new Coordinate((double)1.0, 0, Coordinates.NULL_ORDINATE), new Coordinate((double)0.0, 0,
+        Coordinates.NULL_ORDINATE), new Coordinate((double)1.0, 1,
         Coordinates.NULL_ORDINATE)), 0.000001);
   }
